# docs/swagger.yaml
definitions:
...
  articleColumn.SaveForm:
    properties:
      coverImg:
        type: string
      introduction:
        type: string
      name:
        type: string
    required:
    - coverImg
    - introduction
    - name
    type: object
...
  utils.EmptyInfo:
    type: object
  utils.ResponseResultInfo:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact:
    name: API Support
paths:
...
  /articleColumn/delete:
    delete:
      consumes:
      - application/json
      description: 根据id删除专栏
      parameters:
      - description: 专栏id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseResultInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.EmptyInfo'
      security:
      - ApiKeyAuth: []
      summary: 根据id删除专栏
      tags:
      - articleColumn专栏
  /articleColumn/details:
    get:
      consumes:
      - application/json
      description: 获取专栏详情
      parameters:
      - description: 专栏id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseResultInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.EmptyInfo'
      security:
      - ApiKeyAuth: []
      summary: 获取专栏详情
      tags:
      - articleColumn专栏
  /articleColumn/getAllArticleColumnList:
    get:
      consumes:
      - application/json
      description: 获取全部专栏
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseResultInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.EmptyInfo'
      security:
      - ApiKeyAuth: []
      summary: 获取全部专栏
      tags:
      - articleColumn专栏
  /articleColumn/save:
    post:
      consumes:
      - application/json
      description: 保存专栏
      parameters:
      - description: 请求对象
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/articleColumn.SaveForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseResultInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.EmptyInfo'
      security:
      - ApiKeyAuth: []
      summary: 保存专栏
      tags:
      - articleColumn专栏
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"